Big revert.  Here lies lesson one: Fuck not with that ye dost not fully understand.  I tried to preserve changes which did not break things, but several useful things (signed int to unsigned int conversions) were lost. That would make lesson two: Commit thy cleanups seperately from thy changes if thou wishest not to lose these when thy bugs are reverted.
I've spent several hours reverting this mess by hand and I'm not happy
about it.  That makes lesson three: Maketh not thy commits too large lest
Knghtbrd need to kill thee.  As much as Zeph will not want to hear it, I'm
going to suggest that the interesting parts of his changes be reapplied
carefully.  As for the rest, I think we have found his implementation to
not work right.  I suggest we start over on this front, AFTER 0.2 is out.

+++ trunk/twilight/src/nq/snd_dma.c
@@ -387,7 +387,7 @@ SND_PickChannel (int entnum, int entchannel)
 			break;
 		}
 		// don't let monster sounds override player sounds
-		if (channels[ch_idx].entnum == (int) cl.viewentity && entnum != (int) cl.viewentity
+		if (channels[ch_idx].entnum == cl.viewentity && entnum != cl.viewentity
 			&& channels[ch_idx].sfx)
 			continue;
 
@@ -421,7 +421,7 @@ SND_Spatialize (channel_t *ch)
 	sfx_t      *snd;
 
 // anything coming from the view entity will always be full volume
-	if (ch->entnum == (int) cl.viewentity) {
+	if (ch->entnum == cl.viewentity) {
 		ch->leftvol = ch->master_vol;
 		ch->rightvol = ch->master_vol;
 		return;
